+++ src/z80asm/z80pass.c
@@ -13,7 +13,7 @@
 Copyright (C) Gunther Strube, InterLogic 1993-99
 */
 
-/* $Header: /home/dom/z88dk-git/cvs/z88dk/src/z80asm/z80pass.c,v 1.5 2009-09-03 17:54:55 dom Exp $ */
+/* $Header: /home/dom/z88dk-git/cvs/z88dk/src/z80asm/z80pass.c,v 1.6 2010-04-16 17:34:37 dom Exp $ */
 /* $History: Z80PASS.C $ */
 /*  */
 /* *****************  Version 14  ***************** */
@@ -80,7 +80,7 @@ void WriteLong (long fptr, FILE * fileid);
 void LinkModules (void);
 void DefineOrigin (void);
 void ParseIdent (enum flag interpret);
-void ReportError (char *filename, short linenr, int errnum);
+void ReportError (char *filename, int linenr, int errnum);
 void RemovePfixlist (struct expr *pfixexpr);
 void StoreExpr (struct expr *pfixexpr, char range);
 int DefineSymbol (char *identifier, long value, unsigned char symboltype);
